+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## [v2603.0.0]
 ## Fixed
 - an issue regading passing the storage_options to aiohttp
+### Added
+- Cache cleanup strategy to evict automatically
 
 ## [v2602.1.0]
 ### Added